Update clean MSV

+++ tools/libtiff-test.py
@@ -186,14 +186,15 @@ def num2testcase( case ):
 import psutil
 
 def run_test(testcase,id,env,timeout):
-    test_script=open(testcase,'r')
-    lines=test_script.readlines()
-    test_script.close()
-    if lines[0]=='#!/bin/sh\n':
-        lines[0]='#!/bin/bash\n'
-    test_script=open(testcase,'w')
-    test_script.writelines(lines)
-    test_script.close()
+    if '.sh' in testcase:
+        test_script=open(testcase,'r')
+        lines=test_script.readlines()
+        test_script.close()
+        if lines[0]=='#!/bin/sh\n':
+            lines[0]='#!/bin/bash\n'
+        test_script=open(testcase,'w')
+        test_script.writelines(lines)
+        test_script.close()
 
     proc = subprocess.Popen(["make", "check", "TESTS="+testcase],env = env,stdout=subprocess.PIPE,stderr=subprocess.PIPE)
     try:
@@ -244,8 +245,10 @@ if __name__ == "__main__":
 
         chdir(cur_dir + "/"+temp_dir);
         system("rm -rf o-*.tiff o-*.ps o-*.pdf")
-        system('mv common.sh common-bak.sh')
-        system(f"cp -f {environ['MSV_PATH']}/tools/libtiff-common.sh ./common.sh")
+
+        if 'MSV_PATH' in environ.keys():
+            system('mv common.sh common-bak.sh')
+            system(f"cp -f {environ['MSV_PATH']}/tools/libtiff-common.sh ./common.sh")
 
         my_env = environ;
         my_env["GENEXPOUT"] = "0";
@@ -261,9 +264,10 @@ if __name__ == "__main__":
         pool.close()
         pool.join()
 
-        system('rm -f common.sh')
+        if 'MSV_PATH' in environ.keys():
+            system('rm -f common.sh')
+            system('mv common-bak.sh common.sh')
         system(f'killall --wait {cur_dir}/* > /dev/null 2>&1')
-        system('mv common-bak.sh common.sh')
         chdir(ori_dir);
         subprocess.call('rm -rf '+cur_dir+'/'+temp_dir,shell=True)
